fix: run_sft CKPTS_HOME

+++ tests/e2e/sft/run_sft.sh
@@ -48,4 +48,4 @@ EOF
 
 eval "$cmd"
 
-rm -rf "${CKPTS_HOME}"
+rm -rf "${CKPTS_HOME:?}/*"
